pyproject.toml
[project]
name = "fds"
version = "0.1.0"
description = "Fall Detection System for elderly home care"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.8.0",
    "ultralytics>=8.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "numpy>=1.26.0",
    "apscheduler>=3.11.2",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.14.0",
]

[project.scripts]
fds = "main:main"
fds-test-video = "scripts.test_with_video:main"
fds-cleanup = "scripts.cleanup_clips:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
